requirements.txt
pandas==2.3.2
pyodbc==5.2.0
sqlalchemy==2.0.43
requests
bs4
selenium==4.35.0
webdriver-manager==4.0.2
openpyxl==3.1.5
pysmb==1.2.13
